src/pyproject.toml
[tool.poetry]
name = "companies-scraper"
version = "0.1.0"
description = ""
authors = ["Carlosbogo <bogomcar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.24.0"
google-cloud-bigquery = "^3.25.0"
google-cloud-storage = "^2.18.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
